libs/cloud/requirements.txt
google-cloud==0.32.0
